backend/requirements.txt
python-docx==1.1.0
lxml==4.9.3
PyPDF2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.8
nltk==3.8.1
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
langchain==0.0.350
langchain-community==0.0.10
tiktoken==0.5.1
numpy==1.24.3
scikit-learn==1.3.2
pandas==2.1.4

